# switch to new meta device init method