--- a/fms_fsdp/policies/__init__.py
+++ b/fms_fsdp/policies/__init__.py
@@ -1,3 +1,4 @@
 from .ac_handler import apply_fsdp_checkpointing
 from .mixed_precision import *
+from .param_init import param_init_function
 from .wrapping import get_llama_wrapper