This PR makes a new and improved version of model init which is more efficient.

Full details: https://github.com/foundation-model-stack/fms-fsdp/issues/64

### Performance on model init time (model creation + fsdp wrapping)
We have benchmarked the model init performance on 16 aws p4de nodes:
|   | old implementation |new implementation |
| ------------- | ------------- |------------- |
| 7b  | 100s  | 0.2s |
| 13b | 193s  | 0.3s |
| 34b |  550s  |0.6s |
| 70b |  1146s | 1.5s |

### Validation test
Since unittest is hard to capture true-multi-node runs, we have ran external tests to make sure the new implementation pass validation test and yields true init. Specifically, we added the following plugin after FSDP call for testing purpose
```python
with FSDP.state_dict_type(model, StateDictType.FULL_STATE_DICT, FullStateDictConfig(offload_to_cpu=True, rank0_only=True)):
    model_state = model.state_dict()

if rank == 0:
    model2 = LLaMA(llama_config)
    model2.load_state_dict(model_state)

    model2.check_parameters()
```
and `check_parameters()` is defined [here](https://github.com/foundation-model-stack/foundation-model-stack/commit/8b287887be4d7d668a304ee5dd3555c1f6ad849a)