@lchu-ibm 2024-03-27
Although - that function we added as a one-time thing might be potentially useful to add in FMS too @daviswer .   And it does not depend on anything else.  If we do so, @daviswer just remember to check in your first version (the one I linked above, without the later "wrong fix" like adding `.data`)